feat : SubwayMap duration, distance로 추상화
